Remove invalid optimization from LEEF/FLOWER buffer handling.
Dividing the volume by the number of segments represents was supposed
to represent optimization LEEF can make. Because cells can cover multiple
segments, the data load could be divided among them.

However, this is invalid. While it may be true that we can divide the load
among the cell's segments, the MDC that services that cell will still have to
handle the full volume.