Remove unused dependency on quantities
Older versions of the simulator used the "quantities" package
to verify that we weren't doing anything wierd with units. This
was a great help during early development, but ultimately was
removed due to performance issues. However, there were some
lingering imports for the quantities package. This commit just
removes those.